Fix for bug introduced with the ShadingSystemImpl PIMPL refactor.
Forgot this spot in SSI::attribute(), where it cast a ShadingSystemImpl*
to a ShadingSystem*, back when one was a subclass of the other. Now they
don't have a parent/child relationship. Yikes!

Because of the way optparser calls the system, I needed to add a couple
new overloaded varieties of attribute() to ShadingSystemImp.