Bug fix with instance merging -- never merge the group entry layer.
It turns out that it's a bad idea to do "instance merging" of the last
layer, also known as the "group entry" layer. It's very rare that the
circumstances would lead to it in any real world shader groups, but I
ran across it in a contrived layered example (with two layers of the
same shader master, no connections, same instance parameter values, no
outputs used) The bottom line is that the group entry layer must
generate a proper LLVM function, and having it marked as "unused"
because of merging leads to a cascade of hard-to-understand errors.