oslc fix: allow comma operator
Recently noticed that the "comma operator" does not work as expected.
You can see this with a common idiom of incrementing multiple counters
in a for loop, like this:
    for (int i=0; i < j; ++i, a *= 2) ...

It's not specific to 'for', of course; in general the comma operator
should work in the same places it would in C. To do this properly, we
need to split the grammatical 'expression' into the simple kind and a
compound_expression, and also add a new AST node for the comma operator.