Switch most llvm_opt.cpp's contents to not be externally-visible symbols.
With MCJIT in particular, we can incur a lot of overhead optimizing and JITing
all the multitude of functions compiled into llvm_ops.cpp. This is because they
are all 'extern "C"' which make them look like they might be externally visible,
and therefore potentially callable. But we know this is not the case -- we will
only end up calling our main entry points.

This patch adds functionality to combat this, by iterating over the functions
in the module we've assembled (the vast majority of which should not be fully
compiled if they are not used within the module), and changing their linkage
from ExternalLinkage to LinkOnceODRLinkage. This keeps a single copy that can
be used for inlining or other internal use, but allows it to be discarded if
not used.