OIIO image reader plugin that runs OSL shaders to make the image.
OIIO (and any app that uses it to read images) not only can read many
image formats, but if it doesn't know a format internally, it will look
for a plug-in that can read the format. But it's just a black box, it
doesn't HAVE to actually read an image from disk. It could do something
procedural.

So... I wrote an OIIO image reading plugin that executes OSL code to
procedurally generate image pixels as needed. It can do some fun tricks.

Generate an image by running a compiled OSL shader:

    shader ramp (
        color topleft     = color(0,0,0),
        color topright    = color(1,0,0),
        color bottomleft  = color(0,1,0),
        color bottomright = color(1,1,1),
        output color result = 0,
        output float alpha = 1
        )
    {
        result = topleft*(1-u)*(1-v) + topright*u*(1-v) +
                 bottomleft*(1-u)*v + bottomright*u*v;
    }

    $ oslc ramp.osl
    $ oiiotool ramp.oso -o ramp.exr

You can generate directly into an iv window (or any other app that
uses OIIO to read):

    iv ramp.oso

A ".oslgroup" file can contain the serialization of an entire shading
network:

    oiiotool network.oslgroup -o complex.exr

URI notation can be used to set shader parameters:

    oiiotool "fBm.oso?frequency=4&octaves=4&point offset=0,1,0" -o noise.exr

URI notation also to control the resolution of the "image", or to make
it look tiled rather than scanline, or even to make it behave like it's
MIP-mapped:

    iv "fBm.oso?octaves=4?RES=512x512&MIP=1&TILE=64x64"

".oslbody" signifies a snipped of OSL that will be pasted into
OSL shader boilerplate, compiled from memory, and JITed on the fly:

    iv "result = noise (u*5, v*5);.oslbody"

Not just a single expression, you could write a whole program if you
want:

    iv "int octaves=8; float freq=2,amp=0.5; for (int i = 0; i < octaves; ++i, amp /=2, freq*=2)result += amp*(color)snoise(freq*u,freq*v); result += 0.5; .oslbody?RES=512x512"

Yeah, crazy.