Optimize away unused optional arguments to noise.
Suppose you have a call to noise:

    C = noise (noisetype, P, "bandwidth", val);

Let's suppose that noisetype is a shader parameter. We're passing the
optional "bandwidth" paramter just in case noisetype is "gabor" at
runtime. But if it's not (let's say it's "perlin"), we're needlessly
computing whatever is in 'val'. In fact, "gabor" is currently the only
noise type that uses any of the optional token/value arguments to noise.

This patch culls any optional noise arguments during runtime
optimization, if the noise type can be deduced to be anything other than
"gabor".