Execution order refactor, and per-layer explicit execution order
Let's review what happens when you call

    shadingsys->execute (contextptr, shadergroup, shaderglobals);

This executes the given shader group on the context, with a specific set
of shader globals. The group may be composed of many individual layers
(shader nodes), and execution proceeds from first to last layer,
executing those that must be run unconditionally and skipping those that
can be run lazily. Of course, if a layer executes, it may trigger lazy
execution of other not-yet-run layers. The unconditional layers are the
layers that set global variables (like Ci or P), the layers whose
outputs are connected to designated renderer outputs (AOVs and the
like), and layers that have no connected outputs at all (under the
theory that this must be the "root" node).

This all remains the same with this patch, but *additionally* there is
another option where you can explicitly declare the "entry points" of
a shader group,

    std::vector<const char *> entrynames;  // fill with layer names
    shadingsys->attribute (shadergroup, "entry_layers",
                           TypeDesc(TypeDesc::STRING, nentries),
                           &entrynames[0]);

and then execute the group layer by layer:

    shadingsys->execute_init (contextptr, shadergroup, shaderglobals);
    shadingsys->execute_layer (contextptr, shaderglobals, layernumber);
    ...
    shadingsys->execute_cleanup (contextptr);

The execute_init binds and prepares the context, execute_layer will run
each layer (identified by number, and must correspond to one of the
declared entry layers of the group). As before, other layers will
execute lazily when their results are "pulled" by connections. However,
if you are going to take the approach of manually calling layers, it's
all up to you -- no layers (whose results aren't pulled downstream) will
run unconditionally, even if they appear to set global variables or
appear to be connected to renderer outputs.

In addition to execute_layer by the layer number, there are also similar
calls to specificy the layer by name, and by ShaderSymbol* (running
whichever layer produces the symbol as an output).

The "all in one shot" approach still works just fine, and is probably
the usual way of doing things. But for renderer that need even more
control over which layers are executed and when, this provides an
alternate calling sequence that gives more flexibility.

Some things that had to be done along the way:

* It used to be that the "last" layer was the presumed sole group entry
  point (or "root"), and its code was the one that did certain
  initialization and the calling of the unconditionally-executed layers.
  But now, the app can call any (declared entry) layer, in any order, so
  we can't count on that initialization working. Instead, we now generate
  an additional initialization function, which is called by execute_init.

* For lazy evaluation, it used to be that the *caller* of an earlier layer
  was responsible for marking it as run (so it wouldn't be run twice).
  But now nodes can get called by the app, so the "this layer ran" flag
  is set by the layer itself rather than the caller.

* For the reason immediately above, no layer's LLVM IR code is truly
  empty; at the vest least, it marks the "this layer ran" flag. So the
  prior test for "this whole group does nothing" no longer can count on
  simply checking the post-LLVM optimized IR for being empty but for a
  return statement. This forced us to move the 'group does nothing'
  detection entirely out of LLVM generation and into the time of our
  runtime optimization.

* Groups must track their declared list of possible externally-called
  entry points.

* Augment testshade to be able to declare and call explicit entry
  layers.

* New test: testsuite/layers-entry tests both the default and explicit
  layer execution models, with a test carefully designed to have
  different layer execution orders for the two calling strategies.